dev: renamed proxy to api
