+++ apiserver/plane/api/views/base.py
@@ -17,8 +17,8 @@ from rest_framework import status
 from sentry_sdk import capture_exception
 
 # Module imports
-from plane.proxy.middleware.api_authentication import APIKeyAuthentication
-from plane.proxy.rate_limit import ApiKeyRateThrottle
+from plane.api.middleware.api_authentication import APIKeyAuthentication
+from plane.api.rate_limit import ApiKeyRateThrottle
 from plane.utils.paginator import BasePaginator
 from plane.bgtasks.webhook_task import send_webhook
