+++ apiserver/plane/api/views/inbox.py
@@ -13,7 +13,7 @@ from rest_framework.response import Response
 # Module imports
 from .base import BaseAPIView
 from plane.app.permissions import ProjectLitePermission
-from plane.proxy.serializers import InboxIssueSerializer, IssueSerializer
+from plane.api.serializers import InboxIssueSerializer, IssueSerializer
 from plane.db.models import InboxIssue, Issue, State, ProjectMember
 from plane.bgtasks.issue_activites_task import issue_activity
