--- apiserver/plane/proxy/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ProxyConfig(AppConfig):
-    name = "plane.proxy"
